Moved project in github
I moved my files in GitHub. I have 4 pages with individual CSS files where I have done the menu with little bit of JavaScript so that it is mobile friendly.
Only have little bit of info on the main page as well as a picture of the owner of the portfolio